third_party/WebKit/LayoutTests/fast/forms/select/select-max-length.html

<p>This test that setting HTMLSelectElement.length is capped to 2,147,483,647, and you can't add additional Option elements too.</p>
<div id="console"></div>
<select id="theSelect"></select>
<script src="../../../resources/js-test.js"></script>
<script>
    var sel = document.getElementById('theSelect');
    shouldBe('sel.length', '0');

    debug('Trying: - sel.length = 2147483648;');
    sel.length = 2147483648;
    shouldBe('sel.length', '0');

    debug('Trying: - sel.add(new Option, 0);');
    sel.add(new Option, 0);
    shouldBe('sel.length', '1');

    debug('Trying: - sel.length = 0;');
    sel.length = 0;
    shouldBe('sel.length', '0');

    debug('Index setter:');
    shouldBe('sel[2147483648] = new Option(); sel.length', '0');
    shouldBe('sel.options[2147483648] = new Option(); sel.length', '0');
</script>
